Add molecular dynamics example
