+++ source-code/serial/README.md
@@ -9,6 +9,6 @@ for exercises or for benchmarking.
 
 1. `vector_product_cpu.c`: vector multiplication example.
 1. `julia.c`: application that computes the Julia set.
-1. `molecular_dynamics.cpp`: application that performs a (very naive)
+1. `molecular_dynamics`: application that performs a (very naive)
    molecular dynamics computation.
 1. `CMakeLists.txt`: CMake file to build the applications.
